Move useTranslation to its own file, define TranslationContext directly in _app
