+++ pages/_app.tsx
@@ -1,6 +1,10 @@
 import "globals.css";
 import type { AppProps } from "next/app";
-import { TranslationContext } from "TranslationContext";
+import { createContext } from "react";
+
+export const TranslationContext = createContext<{ translation: Translation }>({
+  translation: {},
+});
 
 function MyApp({ Component, pageProps, router }: AppProps) {
   const translation = require(`../lang/${router.locale}.json`);
